refactor(api): extract swarm routes to src/api/swarm_routes.py [Patch 7/9]
Move 7 swarm routes and the SwarmRuntime singleton getter from
api_server.py into a self-contained route module following the PR #375
pattern.

- swarm_routes.py defines _get_swarm_runtime() locally with adjusted
  swarm_dir path (3-level parent resolution for agent/.swarm/runs)
- register_swarm_routes() resolves require_auth,
  require_event_stream_auth, _validate_path_param, and
  _shell_tools_enabled_for_request via sys.modules fallback
- _swarm_runtime state moves to swarm_routes.py
- Re-export: _get_swarm_runtime for test monkeypatch compat
- api_server.py: -162 lines (2665 -> 2503)

AI-Model: idealab/Qwen3.7-Max-DogFooding
AI-Contributed/Feature: 412/412
AI-Contributed/UT: 0/0
