+++ agent/api_server.py
@@ -1591,179 +1591,17 @@ from src.api.channels_routes import (  # noqa: F401, E402
 )
 
 
+
 # ============================================================================
-# Swarm API
+# Swarm routes - defined in src/api/swarm_routes.py
 # ============================================================================
 
-_swarm_runtime = None
+from src.api.swarm_routes import register_swarm_routes  # noqa: E402
 
+register_swarm_routes(app)
 
-def _get_swarm_runtime():
-    """Lazy-init SwarmRuntime singleton."""
-    global _swarm_runtime
-    if _swarm_runtime is not None:
-        return _swarm_runtime
-    from src.config import load_swarm_agent_config
-    from src.swarm.store import SwarmStore
-    from src.swarm.runtime import SwarmRuntime
-    swarm_dir = Path(__file__).resolve().parent / ".swarm" / "runs"
-    store = SwarmStore(base_dir=swarm_dir)
-    # Boot-time / operator-trusted: REST API callers cannot influence the
-    # config path. See docs/2026-05-25_swarm_mcp_tools_roadmap.md.
-    agent_config = load_swarm_agent_config()
-    _swarm_runtime = SwarmRuntime(store=store, agent_config=agent_config)
-    return _swarm_runtime
-
-
-@app.get("/swarm/presets")
-async def list_swarm_presets():
-    """List Swarm YAML presets."""
-    from src.swarm.presets import list_presets
-    return list_presets()
-
-
-@app.post("/swarm/runs", dependencies=[Depends(require_auth)])
-async def create_swarm_run(payload: dict, http_request: Request):
-    """Start a swarm run: body must include preset_name and user_vars."""
-    runtime = _get_swarm_runtime()
-    preset_name = payload.get("preset_name", "")
-    user_vars = payload.get("user_vars", {})
-    try:
-        run = runtime.start_run(
-            preset_name,
-            user_vars,
-            include_shell_tools=_shell_tools_enabled_for_request(http_request),
-        )
-        return {"id": run.id, "status": run.status.value, "preset_name": run.preset_name}
-    except FileNotFoundError as e:
-        raise HTTPException(status_code=404, detail=str(e))
-    except ValueError as e:
-        raise HTTPException(status_code=400, detail=str(e))
-
-
-@app.get("/swarm/runs", dependencies=[Depends(require_auth)])
-async def list_swarm_runs(limit: int = Query(20, ge=1, le=100)):
-    """List swarm runs (newest first), reconciled."""
-    runtime = _get_swarm_runtime()
-    runs = runtime._store.list_runs(limit=limit)
-    items = []
-    for r in runs:
-        # Reconcile each row: a zombie running run will be auto-finalized so
-        # the dashboard never shows a permanent "running" stuck row.
-        reconciled = runtime._store.reconcile_run(r, write=True)
-        items.append(
-            {
-                "id": reconciled.id,
-                "preset_name": reconciled.preset_name,
-                "status": reconciled.status.value,
-                "is_stale": runtime._store.is_run_stale(reconciled),
-                "created_at": reconciled.created_at,
-                "completed_at": reconciled.completed_at,
-                "task_count": len(reconciled.tasks),
-                "completed_count": sum(1 for t in reconciled.tasks if t.status.value == "completed"),
-            }
-        )
-    return items
-
-
-@app.get("/swarm/runs/{run_id}", dependencies=[Depends(require_auth)])
-async def get_swarm_run(run_id: str):
-    """Swarm run detail including task statuses (reconciled)."""
-    _validate_path_param(run_id, "run_id")
-    runtime = _get_swarm_runtime()
-    loaded = runtime._store.load_run(run_id)
-    if not loaded:
-        raise HTTPException(status_code=404, detail=f"Run {run_id} not found")
-
-    run = runtime._store.reconcile_run(loaded, write=True)
-
-    return {
-        "id": run.id,
-        "preset_name": run.preset_name,
-        "status": run.status.value,
-        "is_stale": runtime._store.is_run_stale(run),
-        "user_vars": run.user_vars,
-        "agents": [a.model_dump() for a in run.agents],
-        "tasks": [t.model_dump() for t in run.tasks],
-        "created_at": run.created_at,
-        "completed_at": run.completed_at,
-        "final_report": run.final_report,
-    }
-
-
-@app.get("/swarm/runs/{run_id}/events", dependencies=[Depends(require_event_stream_auth)])
-async def swarm_run_events(run_id: str, request: Request, last_index: int = Query(0, ge=0)):
-    """SSE stream for a swarm run."""
-    import asyncio
-
-    _validate_path_param(run_id, "run_id")
-    runtime = _get_swarm_runtime()
-
-    async def event_stream():
-        idx = last_index
-        while True:
-            if await request.is_disconnected():
-                break
-            events = runtime._store.read_events(run_id, after_index=idx)
-            for evt in events:
-                idx += 1
-                yield f"id: {idx}\nevent: {evt.type}\ndata: {json.dumps(evt.model_dump(), ensure_ascii=False)}\n\n"
-            run = runtime._store.load_run(run_id)
-            if run:
-                # Reconcile so a zombie running run can still close this SSE
-                # stream cleanly — without it, a dead host would keep the
-                # stream open forever and block the dashboard's "done" state.
-                reconciled = runtime._store.reconcile_run(run, write=True)
-                if reconciled.status.value in ("completed", "failed", "cancelled"):
-                    yield f"event: done\ndata: {{\"status\": \"{reconciled.status.value}\"}}\n\n"
-                    break
-            await asyncio.sleep(2)
-
-    return StreamingResponse(event_stream(), media_type="text/event-stream")
-
-
-@app.post("/swarm/runs/{run_id}/cancel", dependencies=[Depends(require_auth)])
-async def cancel_swarm_run(run_id: str):
-    """Cancel an active swarm run."""
-    _validate_path_param(run_id, "run_id")
-    runtime = _get_swarm_runtime()
-    ok = runtime.cancel_run(run_id)
-    if not ok:
-        raise HTTPException(status_code=404, detail=f"No active run {run_id}")
-    return {"status": "cancelled"}
-
-
-@app.post("/swarm/runs/{run_id}/retry", dependencies=[Depends(require_auth)])
-async def retry_swarm_run(run_id: str, http_request: Request):
-    """Retry a failed, stale, or cancelled swarm run.
-
-    Creates a new run with the same preset and user_vars as the original.
-    """
-    _validate_path_param(run_id, "run_id")
-    runtime = _get_swarm_runtime()
-    loaded = runtime._store.load_run(run_id)
-    if not loaded:
-        raise HTTPException(status_code=404, detail=f"Run {run_id} not found")
-
-    # Reconcile first so a stale "running" run whose host died gets demoted
-    # before we gate on status; only a genuinely active run blocks retry.
-    from src.swarm.models import RunStatus
-
-    reconciled = runtime._store.reconcile_run(loaded, write=True)
-    if reconciled.status == RunStatus.running:
-        raise HTTPException(status_code=409, detail="Cannot retry a running run. Cancel it first.")
-
-    try:
-        new_run = runtime.start_run(
-            reconciled.preset_name,
-            reconciled.user_vars or {},
-            include_shell_tools=_shell_tools_enabled_for_request(http_request),
-        )
-        return {"id": new_run.id, "status": new_run.status.value, "preset_name": new_run.preset_name}
-    except FileNotFoundError as e:
-        raise HTTPException(status_code=404, detail=str(e))
-    except ValueError as e:
-        raise HTTPException(status_code=400, detail=str(e))
+# Re-export for test monkeypatch compatibility
+from src.api.swarm_routes import _get_swarm_runtime  # noqa: F401, E402
 
 
 # ============================================================================
